rename embedding_utils to embeddings
