+++ sliceguard/embeddings.py
